+++ repository/doc.go
@@ -1,17 +1,13 @@
-// Package mirror periodically mirrors (bare clones) remote repositories locally.
+// Package repository periodically mirrors (bare clones) remote repository locally.
+// It supports multiple checked out worktrees on different references.
 // The mirror is created with `--mirror=fetch` hence everything in `refs/*` on the remote
 // will be directly mirrored into `refs/*` in the local repository.
-// it can also maintain multiple mirrored checked out worktrees on different references.
 //
 // The implementation borrows heavily from [kubernetes/git-sync].
 // If you want to sync single repository on one reference then you are probably better off
 // with [kubernetes/git-sync], as it provides a lot more customisation.
 // `git-mirror` should be used if multiple mirrored repositories with multiple
 // checked out branches (worktrees) is required.
-//
-// # Usages
-//
-// please see examples below
 //
 // # Logging:
 //
@@ -33,10 +29,10 @@
 //	}))
 //	loggerLevel.Set(levelStrings["trace"])
 //
-//	repos, err := NewRepoPool(conf, logger.With("logger", "git-mirror"), nil)
+//	repos, err := repository.New(repoConf, nil, logger)
 //	if err != nil {
 //		panic(err)
 //	}
 //
 // [kubernetes/git-sync]: https://github.com/kubernetes/git-sync
-package mirror
+package repository
