+++ repository/helper_test.go
@@ -1,4 +1,4 @@
-package mirror
+package repository
 
 import (
 	"log/slog"
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/utilitywarehouse/git-mirror/internal/utils"
 )
 
 func TestIsFullCommitHash(t *testing.T) {
@@ -135,32 +136,6 @@ func Test_dirIsEmpty(t *testing.T) {
 	}
 }
 
-func Test_reCreate(t *testing.T) {
-	tempRoot := t.TempDir()
-
-	// create files
-	dir := filepath.Join(tempRoot, "files")
-	if err := os.Mkdir(dir, 0755); err != nil {
-		t.Fatalf("failed to make a temp subdir: %v", err)
-	}
-	for _, file := range []string{"a", "b", "c"} {
-		path := filepath.Join(dir, file)
-		if err := os.WriteFile(path, []byte{}, 0755); err != nil {
-			t.Fatalf("failed to write a file: %v", err)
-		}
-	}
-
-	if err := reCreate(dir); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	// validate by making sure new dir is empty
-	if empty, err := dirIsEmpty(dir); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	} else if !empty {
-		t.Errorf("expected %q to be deemed empty", tempRoot)
-	}
-}
-
 func Test_publishSymlink_readAbsLink(t *testing.T) {
 	tempRoot := t.TempDir()
 
@@ -183,7 +158,7 @@ func Test_publishSymlink_readAbsLink(t *testing.T) {
 	}
 
 	// read link to confirm target as destination
-	if dest, err := readAbsLink(link); err != nil {
+	if dest, err := utils.ReadAbsLink(link); err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	} else if dest != target {
 		t.Errorf("expected %q to be the destination of the link but its %q ",
@@ -196,7 +171,7 @@ func Test_publishSymlink_readAbsLink(t *testing.T) {
 	}
 
 	// read link to confirm target as destination
-	if dest, err := readAbsLink(link); err != nil {
+	if dest, err := utils.ReadAbsLink(link); err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	} else if dest != target {
 		t.Errorf("expected %q to be the destination of the link but its %q ",
@@ -214,7 +189,7 @@ func Test_publishSymlink_readAbsLink(t *testing.T) {
 	}
 
 	// read link to confirm target2 as destination
-	if dest, err := readAbsLink(link); err != nil {
+	if dest, err := utils.ReadAbsLink(link); err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	} else if dest != target2 {
 		t.Errorf("expected %q to be the destination of the link but its %q ",
@@ -222,7 +197,7 @@ func Test_publishSymlink_readAbsLink(t *testing.T) {
 	}
 
 	// Test non link absolute path.
-	if _, err := readAbsLink("./link"); err == nil {
+	if _, err := utils.ReadAbsLink("./link"); err == nil {
 		t.Errorf("unexpected success for non absolute path dir")
 	}
 }
@@ -281,36 +256,6 @@ func Test_removeDirContentsIf(t *testing.T) {
 		t.Fatalf("failed to read %q : %v", filepath.Join(target, "b2"), err)
 	}
 
-}
-
-func TestSplitAbs(t *testing.T) {
-	tests := []struct {
-		name    string
-		in      string
-		expDir  string
-		expBase string
-	}{
-		{name: "1", in: "", expDir: "", expBase: ""},
-		{name: "2", in: "/", expDir: "/", expBase: ""},
-		{name: "3", in: "//", expDir: "/", expBase: ""},
-		{name: "4", in: "/one", expDir: "/", expBase: "one"},
-		{name: "5", in: "/one/two", expDir: "/one", expBase: "two"},
-		{name: "6", in: "/one/two/", expDir: "/one", expBase: "two"},
-		{name: "7", in: "/one//two", expDir: "/one", expBase: "two"},
-		{name: "8", in: "one/two", expDir: "one", expBase: "two"},
-		{name: "8", in: "one", expDir: "/", expBase: "one"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got1 := splitAbs(tt.in)
-			if got != tt.expDir {
-				t.Errorf("splitAbs() got = %v, want %v", got, tt.expDir)
-			}
-			if got1 != tt.expBase {
-				t.Errorf("splitAbs() got1 = %v, want %v", got1, tt.expBase)
-			}
-		})
-	}
 }
 
 func Test_updatedRefs(t *testing.T) {
